Do not cancel shell init on mcfly init failure
The shell initialiation codes contain "return 1" at the top level.
However, this causes the cancellation of the entire shell
initialization because those codes are supposed to be evaluated as
e.g. `eval "$(mcfly init bash)"`.  The invocation of `return` in the
eval'ed string takes an effect in the caller context of `eval`.