Fix keyseq specification in Bash
The current way of setting the keyseq as « bind "'\C-r': '...'" ≫
appears to be working as expected, but this is accidental.  This is
not the format that Readline supports; the keyseq cannot be enclosed
in single quotes.  This is now working because of Readline's loose
treatment of the keyseqs, but this will easily be broken with a
slightly different key.